requirements.txt
# Core Dependencies
# Python 3.11+ required

# Async Support
aiohttp==3.9.1
aiofiles==23.2.1
websockets==12.0
asyncpg==0.29.0

# Data Processing
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Caching
redis==5.0.1
aiocache==0.12.2

# Message Queue
aio-pika==9.3.1  # RabbitMQ
aiokafka==0.10.0  # Kafka (optional)

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-aiohttp-client==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-prometheus==0.42b0
opentelemetry-exporter-jaeger==1.21.0

# Logging
python-json-logger==2.0.7
structlog==23.2.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Security & Authentication
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
click==8.1.7
rich==13.7.0
tenacity==8.2.3
cachetools==5.3.2

# Time handling
python-dateutil==2.8.2
pytz==2023.3
pendulum==3.0.0

# API/Web (optional for dashboard)
fastapi==0.104.1
uvicorn==0.25.0
python-multipart==0.0.6

# Plotting/Visualization (for reports)
matplotlib==3.8.2
plotly==5.18.0
kaleido==0.2.1  # For plotly static image export

# PDF Generation
reportlab==4.0.8
weasyprint==60.1

# Excel Support
openpyxl==3.1.2
xlsxwriter==3.1.9

# Machine Learning (optional)
scikit-learn==1.3.2
statsmodels==0.14.1

# Development Dependencies
python-decouple==3.8